+++ alacritty_config_derive/src/config_deserialize/de_struct.rs
@@ -43,13 +43,13 @@ pub fn derive_deserialize<T>(
                 let mut config = Self::Value::default();
 
                 // NOTE: This could be used to print unused keys.
-                let mut unused = serde_yaml::Mapping::new();
+                let mut unused = toml::Table::new();
 
-                while let Some((key, value)) = map.next_entry::<String, serde_yaml::Value>()? {
+                while let Some((key, value)) = map.next_entry::<String, toml::Value>()? {
                     match key.as_str() {
                         #match_assignments
                         _ => {
-                            unused.insert(serde_yaml::Value::String(key), value);
+                            unused.insert(key, value);
                         },
                     }
                 }
@@ -109,7 +109,12 @@ fn field_deserializer(field_streams: &mut FieldStreams, field: &Field) -> Result
         match serde::Deserialize::deserialize(value) {
             Ok(value) => config.#ident = value,
             Err(err) => {
-                log::error!(target: #LOG_TARGET, "Config error: {}: {}", #literal, err);
+                log::error!(
+                    target: #LOG_TARGET,
+                    "Config error: {}: {}",
+                    #literal,
+                    err.to_string().trim(),
+                );
             },
         }
     };
@@ -133,7 +138,7 @@ fn field_deserializer(field_streams: &mut FieldStreams, field: &Field) -> Result
 
                 // Create the tokens to deserialize the flattened struct from the unused fields.
                 field_streams.flatten.extend(quote! {
-                    let unused = serde_yaml::Value::Mapping(unused);
+                    let unused = toml::Value::Table(unused);
                     config.#ident = serde::Deserialize::deserialize(unused).unwrap_or_default();
                 });
             },